requirements.txt
fastapi
uvicorn
easyocr
opencv-python
numpy
pillow
python-multipart
